Short to int  (#440)
* short to int

* re-run migrations

+++ prime-dotnet-webapi/Models/Certifications/College.cs
@@ -6,10 +6,10 @@ using Newtonsoft.Json;
 namespace Prime.Models
 {
     [Table("CollegeLookup")]
-    public class College : BaseAuditable, ILookup<short>
+    public class College : BaseAuditable, ILookup<int>
     {
         [Key]
-        public short Code { get; set; }
+        public int Code { get; set; }
 
         [Required]
         public string Name { get; set; }
